Trees Generate and Burn!
