+++ src/main/java/net/manmaed/cookiesplus/Cookies.java
@@ -7,6 +7,7 @@ import net.manmaed.cookiesplus.libs.Reference;
 import net.manmaed.cookiesplus.proxy.GUIProxy;
 import net.manmaed.cookiesplus.tab.CookiesCreativeTab;
 import net.manmaed.cookiesplus.worldgen.WorldGenCrop;
+import net.manmaed.cookiesplus.worldgen.WorldGenTree;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLFingerprintViolationEvent;
@@ -45,15 +46,14 @@ public class Cookies {
     {
         CookieItems.load();
         CookieBlocks.load();
-        //proxy.renderinfo();
+
     }
 
     @Mod.EventHandler
     public void load(FMLInitializationEvent event)
     {
-        /*proxy.renderinfo();*/
         GameRegistry.registerWorldGenerator(WorldGenCrop.INSTANCE, 5);
-        //GameRegistry.registerWorldGenerator(WorldGenTree.INSTANCE, 5);
+        GameRegistry.registerWorldGenerator(WorldGenTree.INSTANCE, 3);
         NetworkRegistry.INSTANCE.registerGuiHandler(instance, new GUIProxy());
 
     }
